Refactor `jqueryResponse` out to separate file
